[level 2] Title: n^2 배열 자르기, Time: 5.36 ms, Memory: 92.2 MB -BaekjoonHub

+++ 프로그래머스/2/87390. n＾2 배열 자르기/n＾2 배열 자르기.java
@@ -4,8 +4,8 @@ class Solution {
     public int[] solution(int n, long left, long right) {
 
         
-        int standardX = (int) (left % (long) n);
-        int standardY = (int) (left / (long) n);
+        int standardX = (int) (left%n);
+        int standardY = (int) (left/n);
         
         long standard = right - left;
         int[] answer = new int[(int)standard+1];
